+++ tests/contrib/__init__.py
@@ -1,5 +1,3 @@
 # Copyright © 2024 Norman Fomferra and contributors
 # Permissions are hereby granted under the terms of the MIT License:
 # https://opensource.org/licenses/MIT.
-
-__version__ = "0.8.0.dev1"