+++ zappend/contrib/__init__.py
@@ -1,0 +1,15 @@
+# Copyright © 2024 Norman Fomferra and contributors
+# Permissions are hereby granted under the terms of the MIT License:
+# https://opensource.org/licenses/MIT.
+
+"""
+This module contributes to zappend's core functionality.
+
+The function signatures in this module are less stable, and their implementations
+are considered experimental. They may also rely on external packages. For more
+information, please refer to the individual function documentation.
+Due to these reasons, this module is excluded from the project's automatic
+coverage analysis.
+"""
+
+from .levels import write_levels